+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `CallableTree::Node::Internal#compose` (experimental)
+
 ## [0.1.1] - 2021-05-27
 
 - Add `CallableTree::Node::Internal#broadcast` (experimental)